+++ traits/EntityQueryTrait.php
@@ -94,14 +94,14 @@ trait EntityQueryTrait
 
     /**
      * Specify page condition.
-     * @param string|int $currentPage It will return all models if it is 'all',
-     * or it will be regarded as current page number if it is integer begun with 0.
-     * @param int $pageSize The sum of models.
+     * @param string|int $pageSize It will return all models if it is 'all',
+     * or it will be regarded as sum of models.
+     * @param int $currentPage The current page number if it is integer begun with 0.
      * @return $this
      */
-    public function page($currentPage = 0, $pageSize = 10)
+    public function page($pageSize = 10, $currentPage = 0)
     {
-        if ($currentPage === static::$pageAll) {
+        if ($pageSize === static::$pageAll) {
             return $this;
         }
 
@@ -110,7 +110,7 @@ trait EntityQueryTrait
             $currentPage = 0;
         }
         $currentPage = (int) $currentPage;
-        if (!is_int($pageSize) || $pageSize < 1) {
+        if (!is_numeric($pageSize) || $pageSize < 1) {
             $pageSize = static::$defaultPageSize;
         }
         $pageSize = (int) $pageSize;